moodsync-backend/requirements.txt
Flask
Flask-PyMongo
PyJWT
Werkzeug
textblob
